fix(op-acceptance-tests): reduce eth amount and add followup comments
