+++ op-acceptance-tests/acceptance-tests.yaml
@@ -19,6 +19,9 @@ gates:
       - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/base
       - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/base/deposit
         timeout: 10m
+      # TODO(infra#401): Re-enable the test when the sysext missing toolset is implemented
+      #- package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/base/withdrawal
+      #  timeout: 10m
 
   - id: holocene
     inherits:
